// apps/mobile/src/theme/tokens.generated.ts
export const motion = {
  "instant": 0,
  "fast": 120,
  "normal": 180,
  "slow": 220
} as const;